:pencil: fixed useScrollReachedImage file name typo & renamed variables
